removed giss fix, added sam0 unicon comment
