+++ esmvalcore/cmor/_fixes/cmip6/sam0_unicon.py
@@ -1,6 +1,6 @@
 """Fixes for SAM0-UNICON model."""
 from ..common import ClFixHybridPressureCoord
-from .giss_e2_1_g import Nbp as BaseNbp
+from ..fix import Fix
 
 
 Cl = ClFixHybridPressureCoord
@@ -12,4 +12,25 @@ Cli = ClFixHybridPressureCoord
 Clw = ClFixHybridPressureCoord
 
 
-Nbp = BaseNbp
+class Nbp(Fix):
+    """Fixes for nbp."""
+
+    def fix_data(self, cube):
+        """Fix data.
+
+        Fixes wrong sign for land surface flux. Tested for v20190323.
+
+        Parameters
+        ----------
+        cube : iris.cube.Cube
+            Input cube.
+
+        Returns
+        -------
+        iris.cube.Cube
+
+        """
+        metadata = cube.metadata
+        cube *= -1
+        cube.metadata = metadata
+        return cube
